% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot3Variables.R
\name{boxplot3Variables}
\alias{boxplot3Variables}
\title{Create an accessible graph which highlights the changes in price depending on price, type, and brand}
\usage{
boxplot3Variables(
  data,
  cat_variable,
  cat_filteredOn,
  cat_variable_x,
  num_variable_y,
  graph_title
)
}
\arguments{
\item{data}{A data frame.}

\item{cat_variable}{Categorical variable in the data frame to filter on.}

\item{cat_filteredOn}{Filter the categorical variable (cat_variable) according to relevant subcategories.}

\item{cat_variable_x}{Categorical variable to plot on the x-axis of the visualization.}

\item{num_variable_y}{Numerical variable to plot on the y-axis of the visualization.}

\item{graph_title}{The graph's label.}

\item{xaxis_label}{The x-axis's label.}

\item{yaxis_label}{The y-axis's label.}
}
\value{
A boxplot
}
\description{
Create an accessible graph which highlights the changes in price depending on price, type, and brand
}
\examples{
boxplot3Variables(data = shoelati, cat_variable = brand, cat_filteredOn = c("Adidas", "Nike"),
cat_variable_x = type, num_variable_y = price_usd,
graph_title = "Average Price by Type of Shoe According to Brand",
xaxis_label = "Type", yaxis_label = "Price (USD)")

}
