% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emojiViz.R
\name{emojiViz}
\alias{emojiViz}
\title{Create a graph between two numerial variables using emojis as the geom and filtereing by a categorical variable}
\usage{
emojiViz(
  data,
  cat_variable,
  cat_filteredOn,
  num_variable_x,
  num_variable_y,
  emoji_code,
  graph_title,
  xaxis_label,
  yaxis_label,
  num_variable_x_min = 0,
  num_variable_x_max = 100,
  num_variable_y_min = 0,
  num_variable_y_max = 100
)
}
\arguments{
\item{data}{A data frame}

\item{cat_variable}{Categorical variable in the data frame to filter on.}

\item{cat_filteredOn}{Filter the categorical variable (cat_variable) according to relevant subcategories.}

\item{num_variable_x}{Numerical variable to plot on the x-axis of the visualization.}

\item{num_variable_y}{Numerical variable to plot on the y-axis of the visualization.}

\item{emoji_code}{Variable which includes the emoji codes for our dataset.}

\item{graph_title}{The graph's label.}

\item{xaxis_label}{The x-axis's label.}

\item{yaxis_label}{The y-axis's label.}

\item{num_variable_x_min}{The minimum value of the variable plotted on the x-axis.}

\item{num_variable_x_max}{The maximum value of the variable plotted on the x-axis.}

\item{num_variable_y_min}{The minimum value of the variable plotted on the y-axis.}

\item{num_variable_y_max}{The maximum value of the variable plotted on the y-axis.}
}
\value{
A visualization similar to a scatterplot, but which uses emojis as the geom.
}
\description{
Create a graph between two numerial variables using emojis as the geom and filtereing by a categorical variable
}
\examples{
emojiViz(data = shoelati, cat_variable = brand, cat_filteredOn = c("Adidas", "Nike"),
num_variable_x = size_us, num_variable_y = price_usd, emoji_code = emoji,
num_variable_x_min = 6, num_variable_x_max = 12, num_variable_y_min = 25, num_variable_y_max = 250,
graph_title = "Shoe Price According to Shoe Size", xaxis_label = "Size (US)", yaxis_label = "Price (USD)")
}
