% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{summaryFunctions}
\alias{summaryFunctions}
\title{Find summary functions for the data.}
\usage{
summaryFunctions(
  data,
  cat_variable,
  cat_filteredOn,
  num_variable_x,
  num_variable_y,
  title_name
)
}
\arguments{
\item{data}{A data frame}

\item{cat_variable}{Categorical variable in the data frame to filter on.}

\item{cat_filteredOn}{Filter the categorical variable (cat_variable) according to relevant subcategories.}

\item{num_variable_x}{Numerical variable to plot on the x-axis of the visualization.}

\item{num_variable_y}{Numerical variable to plot on the y-axis of the visualization.}
}
\value{
A datatable with the summary functions.
}
\description{
Find summary functions for the data.
}
\examples{
summaryFunctions(data = shoelati, cat_variable = brand, cat_filteredOn = c("Adidas", "Nike"),
num_variable_x = size_us, num_variable_y = price_usd)

}
