% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot3Variables.R
\name{boxplot3Variables}
\alias{boxplot3Variables}
\title{An accessible boxplot according to three variables.}
\usage{
boxplot3Variables(
  data,
  variable_toFilter,
  variable_filteredOn,
  explanatory_variable,
  response_variable,
  graph_title,
  xaxis_label,
  yaxis_label
)
}
\arguments{
\item{data}{A data frame}

\item{variable_toFilter}{One numeric or categorical variable in the data frame to filter on.}

\item{variable_filteredOn}{Filter the numeric or categorical variable (variable_toFilter) according to relevant subcategories of interest.}

\item{explanatory_variable}{One unique numeric or categorical variable to group the data by to better understand the response variable.}

\item{response_variable}{One numeric variable whose values are analyzed in the context of the explanatory and filtered variable to draw trends in the data.}

\item{graph_title}{The graph's label.}

\item{xaxis_label}{The x-axis's label.}

\item{yaxis_label}{The y-axis's label.}
}
\value{
A boxplot highlighting trends according of the variable we have filtered on for by our chosen response and explanatory variables.
}
\description{
Creates an accessible boxplot which highlights the relationship between a numeric variable and two other variables (either numeric or categorical).
}
\examples{
boxplot3Variables(data = shoelati, variable_toFilter = brand,
variable_filteredOn = c("Adidas", "Nike"),
explanatory_variable = type, response_variable = price_usd,
graph_title = "Average Price by Type of Shoe According to Brand",
xaxis_label = "Type", yaxis_label = "Price (USD)")


}
