% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emojiViz.R
\name{emojiViz}
\alias{emojiViz}
\title{An emoji visualization}
\usage{
emojiViz(
  data,
  variable_toFilter,
  variable_filteredOn,
  explanatory_variable,
  response_variable,
  emoji_code,
  graph_title,
  xaxis_label,
  yaxis_label,
  num_variable_x_min = 0,
  num_variable_x_max = 100,
  num_variable_y_min = 0,
  num_variable_y_max = 100
)
}
\arguments{
\item{data}{A data frame}

\item{variable_toFilter}{One numeric or categorical variable in the data frame to filter on.}

\item{variable_filteredOn}{Filter the numeric or categorical variable (variable_toFilter) according to relevant subcategories of interest.}

\item{explanatory_variable}{One NUMERIC variable mapped to the x-axis (this is different from the other functions which can use numeric or categorical inputs).}

\item{response_variable}{One numeric variable to mapped to the y-axis.}

\item{emoji_code}{The unicode for emojis matched to the rows in the dataframe}

\item{graph_title}{The graph's label.}

\item{xaxis_label}{The x-axis's label.}

\item{yaxis_label}{The y-axis's label.}

\item{num_variable_x_min}{The minimum value of the variable plotted on the x-axis.}

\item{num_variable_x_max}{The maximum value of the variable plotted on the x-axis.}

\item{num_variable_y_min}{The minimum value of the variable plotted on the y-axis.}

\item{num_variable_y_max}{The maximum value of the variable plotted on the y-axis.}
}
\value{
A visualization similar to a scatterplot, but which uses emojis as the geom.   Rows may be removed in order to make the visualization readable.
}
\description{
Creates a graph using two numeric variables and emojis as the geom (filtering by a categorical variable) to determine trends in general shoe type and variables of interest.
}
\examples{
emojiViz(data = shoelati, variable_toFilter = brand,
variable_filteredOn = c("Adidas", "Nike"),explanatory_variable = size_us,
response_variable = price_usd, emoji_code = emoji,
num_variable_x_min = 6, num_variable_x_max = 12,
num_variable_y_min = 25, num_variable_y_max = 250,
graph_title = "Shoe Price According to Shoe Size",
xaxis_label = "Size (US)", yaxis_label = "Price (USD)")
}
