% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{summaryFunctions}
\alias{summaryFunctions}
\title{Perform summary functions.}
\usage{
summaryFunctions(
  data,
  variable_toFilter,
  variable_filteredOn,
  explanatory_variable,
  response_variable
)
}
\arguments{
\item{data}{A data frame}

\item{variable_toFilter}{One numeric or categorical variable in the data frame to filter on.}

\item{variable_filteredOn}{Filter the numeric or categorical variable (variable_toFilter) according to relevant subcategories of interest.}

\item{explanatory_variable}{One unique numeric or categorical variable to group the data by to better understand the response variable.}

\item{response_variable}{One numeric variable whose values are analyzed in the context of the explanatory and filtered variable to draw trends in the data.}
}
\value{
A datatable with multiple summary functions perfomed on a chosen response variable.    Dimensions depend on the number of categories we have filtered on and the number of values for our explanatory variable.
}
\description{
Calculates summary functions (mean, median, etc.) for a response variable after grouping or filtering by two specific other variables (either numeric or categorical) to detect trends.
}
\examples{
summaryFunctions(data = shoelati, variable_toFilter = brand,
variable_filteredOn = c("Adidas", "Nike"),
explanatory_variable = size_us, response_variable = price_usd)

}
